third_party/WebKit/LayoutTests/external/wpt/permafill/import-g3.sub.html

<!DOCTYPE html>
<title>Import inside Permafills (Syntax G.3)</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    window.log = [];
</script>

<script type="module" onload="log.push('load')" onerror="log.push('error')"
    src="http://{{host}}:{{ports[http][0]}}/permafill/.well-known/permafill/test2.js"></script>

<script type="module">
    test(() => {
        assert_array_equals(log, [
          "browser://test2/sub.js", "browser://test2/", "load"
        ]);
      }, "Load permafill w/ import by a <script> element");
</script>
